feat : 비행 스케줄 등록 1차
- 추가 해야할 것
  - 비행 스케줄에 가능한 좌석들
    - 좌석에 따른 옵션 (기내식, 수하물, 디스플레이, 와이파이, usb 등)
  - 비행 스케줄 등록 시 공항의 스케줄 고려
    - 공항의 터미널 별 입출국 가능 시간 확인
    - 입출국 가능 시간에 따른 예상 비행 시간 조정

+++ src/main/java/com/example/jpamaster/flight/constant/FlightConstant.java
@@ -1,6 +1,6 @@
 package com.example.jpamaster.flight.constant;
 
 public class FlightConstant {
-    public static final int AVERAGE_AIRPLANE_SPEED = 900;
+    public static final double AVERAGE_AIRPLANE_SPEED = 900.0;
     public static final int EARTH_RADIUS = 6371;
 }
